requirements.txt
python-magic;platform_system != "Windows"
python-magic-bin;platform_system=="Windows"
windnd
